исправление ошибочного перемещения
