+++ students-registry/users-service-test-util/src/main/java/boost/brain/course/users/test/util/DeleteSpam.java
@@ -1,4 +1,4 @@
-package boost.brain.course.projects.controller.util;
+package boost.brain.course.users.test.util;
 
 import boost.brain.course.common.users.UserDto;
 import lombok.extern.java.Log;
@@ -11,19 +11,19 @@ import java.util.List;
 import java.util.Random;
 import java.util.concurrent.RecursiveTask;
 
-import static boost.brain.course.projects.controller.Constants.*;
+import static boost.brain.course.users.test.Constants.*;
 
 @Log
-public class ReadSpam extends RecursiveTask<Boolean> {
+public class DeleteSpam extends RecursiveTask<Boolean> {
     private Long count;
     private String url;
     private RestTemplate restTemplate;
     private Random random = new Random();
 
-    public ReadSpam() {
+    public DeleteSpam() {
     }
 
-    public ReadSpam(String url, Long count, RestTemplate restTemplate) {
+    public DeleteSpam(String url, Long count, RestTemplate restTemplate) {
         this.count = count;
         this.restTemplate = restTemplate;
         this.url = url;
@@ -33,20 +33,20 @@ public class ReadSpam extends RecursiveTask<Boolean> {
     protected Boolean compute() {
         if (count / 2 > 0) {
             if(count % 2 == 1){
-                executeRead();
+                executeDelete();
             }
-            ReadSpam readSpamOne = new ReadSpam(url, count / 2, restTemplate);
-            readSpamOne.fork();
-            ReadSpam readSpamTwo = new ReadSpam(url, count / 2, restTemplate);
-            readSpamTwo.fork();
+            DeleteSpam deleteSpamOne = new DeleteSpam(url, count / 2, restTemplate);
+            deleteSpamOne.fork();
+            DeleteSpam deleteSpamTwo = new DeleteSpam(url, count / 2, restTemplate);
+            deleteSpamTwo.fork();
         } else {
-            executeRead();
+            executeDelete();
             return  true;
         }
         return  true;
     }
 
-    private void executeRead() {
+    private void executeDelete() {
         int randomPage = random.nextInt(1000);
         long startTime = System.currentTimeMillis();
 
@@ -63,11 +63,15 @@ public class ReadSpam extends RecursiveTask<Boolean> {
         for (UserDto user: users) {
             email = user.getEmail();
             startTime = System.currentTimeMillis();
-            UserDto readUser = restTemplate.getForObject(url + USERS_CONTROLLER_PREFIX + "/" + READ_PREFIX + "/" + email, UserDto.class);
+            try {
+                restTemplate.delete(url + USERS_CONTROLLER_PREFIX + "/" + DELETE_PREFIX + "/" + email);
+            }catch (Exception e){
+                log.info(e.getLocalizedMessage());
+            }
 
             stopTime = System.currentTimeMillis();
             elapsedTime = stopTime - startTime;
-            log.info("Read " + readUser.getEmail() + " time :" + elapsedTime + "ms " );
+            log.info("Delete " + email + " time :" + elapsedTime + "ms " );
         }
     }
