+++ students-registry/users-service-test-util/src/main/java/boost/brain/course/users/test/util/UpdateStatusesForEmailsSpam.java
@@ -1,4 +1,4 @@
-package boost.brain.course.projects.controller.util;
+package boost.brain.course.users.test.util;
 import boost.brain.course.common.users.UserDto;
 import boost.brain.course.common.users.UserStatus;
 import lombok.extern.java.Log;
@@ -12,7 +12,7 @@ import org.springframework.web.client.RestTemplate;
 import java.util.*;
 import java.util.concurrent.RecursiveTask;
 
-import static boost.brain.course.projects.controller.Constants.*;
+import static boost.brain.course.users.test.Constants.*;
 
 @Log
 public class UpdateStatusesForEmailsSpam extends RecursiveTask<Boolean> {
